Quarter-2/Assignment-03-openai-agent-sdk/finadvisor/pyproject.toml
[project]
name = "finadvisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Eshban Bahadur", email = "eshban@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]

[project.scripts]
finadvisor = "finadvisor:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
